Improved HLPP

+++ content/graph/PushRelabel.h
@@ -18,8 +18,8 @@ template<typename Flow = lint> struct PushRelabel {
 	PushRelabel(int n) : g(n), ec(n), cur(n), hs(2*n), H(n) {}
 	void addEdge(int s, int t, Flow cap, Flow rcap = 0) {
 		if (s == t) return;
-		g[s].push_back({t, g[t].size(), 0, cap});
-		g[t].push_back({s, g[s].size()-1, 0, rcap});
+		g[s].push_back({t, (int)g[t].size(), 0, cap});
+		g[t].push_back({s, (int)g[s].size()-1, 0, rcap});
 	}
 	void addFlow(edge_t& e, Flow f) {
 		edge_t &back = g[e.dest][e.back];
